+++ src/attr/validators.py
@@ -88,17 +88,19 @@ def provides(interface):
 
 @attributes(repr=False, slots=True)
 class _OptionalValidator(object):
-    validator = attr()
+    validators = attr()
 
     def __call__(self, inst, attr, value):
         if value is None:
             return
-        return self.validator(inst, attr, value)
+
+        for v in self.validators:
+            v(inst, attr, value)
 
     def __repr__(self):
         return (
             "<optional validator for {type} or None>"
-            .format(type=repr(self.validator))
+            .format(type=repr(self.validators))
         )
 
 
@@ -108,6 +110,12 @@ def optional(validator):
     which can be set to ``None`` in addition to satisfying the requirements of
     the sub-validator.
 
-    :param validator: A validator that is used for non-``None`` values.
+    :param validator: A validator (or a list of validators) that is used for
+        non-``None`` values.
+    :type validator: callable or :class:`list` of callables.
+
+    .. versionchanged:: 17.1.0 *validator* can be a list of validators.
     """
-    return _OptionalValidator(validator)
+    if isinstance(validator, list):
+        return _OptionalValidator(tuple(validator))
+    return _OptionalValidator((validator,))