+++ tests/test_validators.py
@@ -111,29 +111,39 @@ class TestProvides(object):
         ) == repr(v)
 
 
+def always_pass(_, __, ___):
+    """
+    Toy validator that always passses.
+    """
+
+
+@pytest.mark.parametrize("validator", [
+    instance_of(int),
+    [always_pass, instance_of(int)],
+])
 class TestOptional(object):
     """
     Tests for `optional`.
     """
-    def test_success_with_type(self):
+    def test_success(self, validator):
         """
-        Nothing happens if types match.
+        Nothing happens if validator succeeds.
         """
-        v = optional(instance_of(int))
+        v = optional(validator)
         v(None, simple_attr("test"), 42)
 
-    def test_success_with_none(self):
+    def test_success_with_none(self, validator):
         """
         Nothing happens if None.
         """
-        v = optional(instance_of(int))
+        v = optional(validator)
         v(None, simple_attr("test"), None)
 
-    def test_fail(self):
+    def test_fail(self, validator):
         """
         Raises `TypeError` on wrong types.
         """
-        v = optional(instance_of(int))
+        v = optional(validator)
         a = simple_attr("test")
         with pytest.raises(TypeError) as e:
             v(None, a, "42")
@@ -144,13 +154,21 @@ class TestOptional(object):
 
         ) == e.value.args
 
-    def test_repr(self):
+    def test_repr(self, validator):
         """
         Returned validator has a useful `__repr__`.
         """
-        v = optional(instance_of(int))
-        assert (
-            ("<optional validator for <instance_of validator for type "
-             "<{type} 'int'>> or None>")
-            .format(type=TYPE)
-        ) == repr(v)
+        v = optional(validator)
+
+        if isinstance(validator, list):
+            assert (
+                ("<optional validator for ({func}, <instance_of validator for "
+                 "type <{type} 'int'>>) or None>")
+                .format(func=repr(always_pass), type=TYPE)
+            ) == repr(v)
+        else:
+            assert (
+                ("<optional validator for (<instance_of validator for type "
+                 "<{type} 'int'>>,) or None>")
+                .format(type=TYPE)
+            ) == repr(v)